requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.14.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
folium>=0.14.0
streamlit-folium>=0.11.0
python-dotenv>=1.0.0
